Site Editor: Fix sidebar plugins
